Changed format of printing to time to a float.  Now has precision of .1 ms.
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@1318 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/io/LinearActivityProbe.cpp
@@ -53,7 +53,7 @@ int LinearActivityProbe::outputState(float time, PVLayer * l)
    }
 
    float freq = sum / (nk * dt * 0.001);
-   fprintf(fp, "t=%4d sum=%3d f=%6.1f Hz :", (int)time, (int)sum, freq);
+   fprintf(fp, "t=%6.1f sum=%3d f=%6.1f Hz :", time, (int)sum, freq);
 
    for (int k = 0; k < nk; k++) {
       float a = line[f + k * sk];
